+++ datafusion/functions-array/src/kernels.rs
@@ -17,29 +17,25 @@
 
 //! implementation kernels for array functions
 
+use arrow::array::ListArray;
 use arrow::array::{
     Array, ArrayRef, BooleanArray, Date32Array, Float32Array, Float64Array,
     GenericListArray, Int16Array, Int32Array, Int64Array, Int8Array, LargeStringArray,
     OffsetSizeTrait, StringArray, UInt16Array, UInt32Array, UInt64Array, UInt8Array,
 };
-use arrow::array::{ArrayData, Capacities, ListArray, MutableArrayData};
 use arrow::buffer::OffsetBuffer;
-use arrow::compute;
 use arrow::datatypes::Field;
 use arrow::datatypes::UInt64Type;
 use arrow::datatypes::{DataType, Date32Type, IntervalMonthDayNanoType};
-use arrow_array::{new_null_array, NullArray};
-use arrow_buffer::{BooleanBufferBuilder, NullBuffer};
 use datafusion_common::cast::{
-    as_date32_array, as_generic_list_array, as_int64_array, as_interval_mdn_array,
-    as_large_list_array, as_list_array, as_string_array,
+    as_date32_array, as_int64_array, as_interval_mdn_array, as_large_list_array,
+    as_list_array, as_string_array,
 };
-use datafusion_common::utils::{array_into_list_array, list_ndims};
-use datafusion_common::{
-    exec_err, not_impl_datafusion_err, not_impl_err, plan_err, DataFusionError, Result,
-};
+use datafusion_common::DataFusionError;
+use datafusion_common::{exec_err, not_impl_datafusion_err, Result};
 use std::any::type_name;
 use std::sync::Arc;
+
 macro_rules! downcast_arg {
     ($ARG:expr, $ARRAY_TYPE:ident) => {{
         $ARG.as_any().downcast_ref::<$ARRAY_TYPE>().ok_or_else(|| {
@@ -522,433 +518,4 @@ pub fn gen_range_date(
         None,
     )?);
     Ok(arr)
-}
-
-/// Appends or prepends elements to a ListArray.
-///
-/// This function takes a ListArray, an ArrayRef, a FieldRef, and a boolean flag
-/// indicating whether to append or prepend the elements. It returns a `Result<ArrayRef>`
-/// representing the resulting ListArray after the operation.
-///
-/// # Arguments
-///
-/// * `list_array` - A reference to the ListArray to which elements will be appended/prepended.
-/// * `element_array` - A reference to the Array containing elements to be appended/prepended.
-/// * `field` - A reference to the Field describing the data type of the arrays.
-/// * `is_append` - A boolean flag indicating whether to append (`true`) or prepend (`false`) elements.
-///
-/// # Examples
-///
-/// generic_append_and_prepend(
-///     [1, 2, 3], 4, append => [1, 2, 3, 4]
-///     5, [6, 7, 8], prepend => [5, 6, 7, 8]
-/// )
-fn generic_append_and_prepend<O: OffsetSizeTrait>(
-    list_array: &GenericListArray<O>,
-    element_array: &ArrayRef,
-    data_type: &DataType,
-    is_append: bool,
-) -> Result<ArrayRef>
-where
-    i64: TryInto<O>,
-{
-    let mut offsets = vec![O::usize_as(0)];
-    let values = list_array.values();
-    let original_data = values.to_data();
-    let element_data = element_array.to_data();
-    let capacity = Capacities::Array(original_data.len() + element_data.len());
-
-    let mut mutable = MutableArrayData::with_capacities(
-        vec![&original_data, &element_data],
-        false,
-        capacity,
-    );
-
-    let values_index = 0;
-    let element_index = 1;
-
-    for (row_index, offset_window) in list_array.offsets().windows(2).enumerate() {
-        let start = offset_window[0].to_usize().unwrap();
-        let end = offset_window[1].to_usize().unwrap();
-        if is_append {
-            mutable.extend(values_index, start, end);
-            mutable.extend(element_index, row_index, row_index + 1);
-        } else {
-            mutable.extend(element_index, row_index, row_index + 1);
-            mutable.extend(values_index, start, end);
-        }
-        offsets.push(offsets[row_index] + O::usize_as(end - start + 1));
-    }
-
-    let data = mutable.freeze();
-
-    Ok(Arc::new(GenericListArray::<O>::try_new(
-        Arc::new(Field::new("item", data_type.to_owned(), true)),
-        OffsetBuffer::new(offsets.into()),
-        arrow_array::make_array(data),
-        None,
-    )?))
-}
-
-fn check_datatypes(name: &str, args: &[&ArrayRef]) -> Result<()> {
-    let data_type = args[0].data_type();
-    if !args.iter().all(|arg| {
-        arg.data_type().equals_datatype(data_type)
-            || arg.data_type().equals_datatype(&DataType::Null)
-    }) {
-        let types = args.iter().map(|arg| arg.data_type()).collect::<Vec<_>>();
-        return plan_err!("{name} received incompatible types: '{types:?}'.");
-    }
-
-    Ok(())
-}
-
-fn general_append_and_prepend<O: OffsetSizeTrait>(
-    args: &[ArrayRef],
-    is_append: bool,
-) -> Result<ArrayRef>
-where
-    i64: TryInto<O>,
-{
-    let (list_array, element_array) = if is_append {
-        let list_array = as_generic_list_array::<O>(&args[0])?;
-        let element_array = &args[1];
-        check_datatypes("array_append", &[element_array, list_array.values()])?;
-        (list_array, element_array)
-    } else {
-        let list_array = as_generic_list_array::<O>(&args[1])?;
-        let element_array = &args[0];
-        check_datatypes("array_prepend", &[list_array.values(), element_array])?;
-        (list_array, element_array)
-    };
-
-    let res = match list_array.value_type() {
-        DataType::List(_) => concat_internal::<i32>(args)?,
-        DataType::LargeList(_) => concat_internal::<i64>(args)?,
-        data_type => {
-            return generic_append_and_prepend::<O>(
-                list_array,
-                element_array,
-                &data_type,
-                is_append,
-            );
-        }
-    };
-
-    Ok(res)
-}
-
-/// Convert one or more [`ArrayRef`] of the same type into a
-/// `ListArray` or 'LargeListArray' depending on the offset size.
-///
-/// # Example (non nested)
-///
-/// Calling `array(col1, col2)` where col1 and col2 are non nested
-/// would return a single new `ListArray`, where each row was a list
-/// of 2 elements:
-///
-/// ```text
-/// ┌─────────┐   ┌─────────┐           ┌──────────────┐
-/// │ ┌─────┐ │   │ ┌─────┐ │           │ ┌──────────┐ │
-/// │ │  A  │ │   │ │  X  │ │           │ │  [A, X]  │ │
-/// │ ├─────┤ │   │ ├─────┤ │           │ ├──────────┤ │
-/// │ │NULL │ │   │ │  Y  │ │──────────▶│ │[NULL, Y] │ │
-/// │ ├─────┤ │   │ ├─────┤ │           │ ├──────────┤ │
-/// │ │  C  │ │   │ │  Z  │ │           │ │  [C, Z]  │ │
-/// │ └─────┘ │   │ └─────┘ │           │ └──────────┘ │
-/// └─────────┘   └─────────┘           └──────────────┘
-///   col1           col2                    output
-/// ```
-///
-/// # Example (nested)
-///
-/// Calling `array(col1, col2)` where col1 and col2 are lists
-/// would return a single new `ListArray`, where each row was a list
-/// of the corresponding elements of col1 and col2.
-///
-/// ``` text
-/// ┌──────────────┐   ┌──────────────┐        ┌─────────────────────────────┐
-/// │ ┌──────────┐ │   │ ┌──────────┐ │        │ ┌────────────────────────┐  │
-/// │ │  [A, X]  │ │   │ │    []    │ │        │ │    [[A, X], []]        │  │
-/// │ ├──────────┤ │   │ ├──────────┤ │        │ ├────────────────────────┤  │
-/// │ │[NULL, Y] │ │   │ │[Q, R, S] │ │───────▶│ │ [[NULL, Y], [Q, R, S]] │  │
-/// │ ├──────────┤ │   │ ├──────────┤ │        │ ├────────────────────────│  │
-/// │ │  [C, Z]  │ │   │ │   NULL   │ │        │ │    [[C, Z], NULL]      │  │
-/// │ └──────────┘ │   │ └──────────┘ │        │ └────────────────────────┘  │
-/// └──────────────┘   └──────────────┘        └─────────────────────────────┘
-///      col1               col2                         output
-/// ```
-fn array_array<O: OffsetSizeTrait>(
-    args: &[ArrayRef],
-    data_type: DataType,
-) -> Result<ArrayRef> {
-    // do not accept 0 arguments.
-    if args.is_empty() {
-        return plan_err!("Array requires at least one argument");
-    }
-
-    let mut data = vec![];
-    let mut total_len = 0;
-    for arg in args {
-        let arg_data = if arg.as_any().is::<NullArray>() {
-            ArrayData::new_empty(&data_type)
-        } else {
-            arg.to_data()
-        };
-        total_len += arg_data.len();
-        data.push(arg_data);
-    }
-
-    let mut offsets: Vec<O> = Vec::with_capacity(total_len);
-    offsets.push(O::usize_as(0));
-
-    let capacity = Capacities::Array(total_len);
-    let data_ref = data.iter().collect::<Vec<_>>();
-    let mut mutable = MutableArrayData::with_capacities(data_ref, true, capacity);
-
-    let num_rows = args[0].len();
-    for row_idx in 0..num_rows {
-        for (arr_idx, arg) in args.iter().enumerate() {
-            if !arg.as_any().is::<NullArray>()
-                && !arg.is_null(row_idx)
-                && arg.is_valid(row_idx)
-            {
-                mutable.extend(arr_idx, row_idx, row_idx + 1);
-            } else {
-                mutable.extend_nulls(1);
-            }
-        }
-        offsets.push(O::usize_as(mutable.len()));
-    }
-    let data = mutable.freeze();
-
-    Ok(Arc::new(GenericListArray::<O>::try_new(
-        Arc::new(Field::new("item", data_type, true)),
-        OffsetBuffer::new(offsets.into()),
-        arrow_array::make_array(data),
-        None,
-    )?))
-}
-
-/// Array_append SQL function
-pub(crate) fn array_append(args: &[ArrayRef]) -> Result<ArrayRef> {
-    if args.len() != 2 {
-        return exec_err!("array_append expects two arguments");
-    }
-
-    match args[0].data_type() {
-        DataType::LargeList(_) => general_append_and_prepend::<i64>(args, true),
-        _ => general_append_and_prepend::<i32>(args, true),
-    }
-}
-
-/// Array_prepend SQL function
-pub fn array_prepend(args: &[ArrayRef]) -> Result<ArrayRef> {
-    if args.len() != 2 {
-        return exec_err!("array_prepend expects two arguments");
-    }
-
-    match args[1].data_type() {
-        DataType::LargeList(_) => general_append_and_prepend::<i64>(args, false),
-        _ => general_append_and_prepend::<i32>(args, false),
-    }
-}
-
-/// Array_concat/Array_cat SQL function
-pub fn array_concat(args: &[ArrayRef]) -> Result<ArrayRef> {
-    if args.is_empty() {
-        return exec_err!("array_concat expects at least one arguments");
-    }
-
-    let mut new_args = vec![];
-    for arg in args {
-        let ndim = list_ndims(arg.data_type());
-        let base_type = datafusion_common::utils::base_type(arg.data_type());
-        if ndim == 0 {
-            return not_impl_err!("Array is not type '{base_type:?}'.");
-        }
-        if !base_type.eq(&DataType::Null) {
-            new_args.push(arg.clone());
-        }
-    }
-
-    match &args[0].data_type() {
-        DataType::LargeList(_) => concat_internal::<i64>(new_args.as_slice()),
-        _ => concat_internal::<i32>(new_args.as_slice()),
-    }
-}
-
-/// `make_array` SQL function
-pub fn make_array(arrays: &[ArrayRef]) -> Result<ArrayRef> {
-    let mut data_type = DataType::Null;
-    for arg in arrays {
-        let arg_data_type = arg.data_type();
-        if !arg_data_type.equals_datatype(&DataType::Null) {
-            data_type = arg_data_type.clone();
-            break;
-        }
-    }
-
-    match data_type {
-        // Either an empty array or all nulls:
-        DataType::Null => {
-            let array =
-                new_null_array(&DataType::Null, arrays.iter().map(|a| a.len()).sum());
-            Ok(Arc::new(array_into_list_array(array)))
-        }
-        DataType::LargeList(..) => array_array::<i64>(arrays, data_type),
-        _ => array_array::<i32>(arrays, data_type),
-    }
-}
-
-fn concat_internal<O: OffsetSizeTrait>(args: &[ArrayRef]) -> Result<ArrayRef> {
-    let args = align_array_dimensions::<O>(args.to_vec())?;
-
-    let list_arrays = args
-        .iter()
-        .map(|arg| as_generic_list_array::<O>(arg))
-        .collect::<Result<Vec<_>>>()?;
-    // Assume number of rows is the same for all arrays
-    let row_count = list_arrays[0].len();
-
-    let mut array_lengths = vec![];
-    let mut arrays = vec![];
-    let mut valid = BooleanBufferBuilder::new(row_count);
-    for i in 0..row_count {
-        let nulls = list_arrays
-            .iter()
-            .map(|arr| arr.is_null(i))
-            .collect::<Vec<_>>();
-
-        // If all the arrays are null, the concatenated array is null
-        let is_null = nulls.iter().all(|&x| x);
-        if is_null {
-            array_lengths.push(0);
-            valid.append(false);
-        } else {
-            // Get all the arrays on i-th row
-            let values = list_arrays
-                .iter()
-                .map(|arr| arr.value(i))
-                .collect::<Vec<_>>();
-
-            let elements = values
-                .iter()
-                .map(|a| a.as_ref())
-                .collect::<Vec<&dyn Array>>();
-
-            // Concatenated array on i-th row
-            let concated_array = compute::concat(elements.as_slice())?;
-            array_lengths.push(concated_array.len());
-            arrays.push(concated_array);
-            valid.append(true);
-        }
-    }
-    // Assume all arrays have the same data type
-    let data_type = list_arrays[0].value_type();
-    let buffer = valid.finish();
-
-    let elements = arrays
-        .iter()
-        .map(|a| a.as_ref())
-        .collect::<Vec<&dyn Array>>();
-
-    let list_arr = GenericListArray::<O>::new(
-        Arc::new(Field::new("item", data_type, true)),
-        OffsetBuffer::from_lengths(array_lengths),
-        Arc::new(compute::concat(elements.as_slice())?),
-        Some(NullBuffer::new(buffer)),
-    );
-
-    Ok(Arc::new(list_arr))
-}
-
-fn align_array_dimensions<O: OffsetSizeTrait>(
-    args: Vec<ArrayRef>,
-) -> Result<Vec<ArrayRef>> {
-    let args_ndim = args
-        .iter()
-        .map(|arg| datafusion_common::utils::list_ndims(arg.data_type()))
-        .collect::<Vec<_>>();
-    let max_ndim = args_ndim.iter().max().unwrap_or(&0);
-
-    // Align the dimensions of the arrays
-    let aligned_args: Result<Vec<ArrayRef>> = args
-        .into_iter()
-        .zip(args_ndim.iter())
-        .map(|(array, ndim)| {
-            if ndim < max_ndim {
-                let mut aligned_array = array.clone();
-                for _ in 0..(max_ndim - ndim) {
-                    let data_type = aligned_array.data_type().to_owned();
-                    let array_lengths = vec![1; aligned_array.len()];
-                    let offsets = OffsetBuffer::<O>::from_lengths(array_lengths);
-
-                    aligned_array = Arc::new(GenericListArray::<O>::try_new(
-                        Arc::new(Field::new("item", data_type, true)),
-                        offsets,
-                        aligned_array,
-                        None,
-                    )?)
-                }
-                Ok(aligned_array)
-            } else {
-                Ok(array.clone())
-            }
-        })
-        .collect();
-
-    aligned_args
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use arrow::datatypes::Int64Type;
-    use arrow_array::ListArray;
-
-    /// Only test internal functions, array-related sql functions will be tested in sqllogictest `array.slt`
-    #[test]
-    fn test_align_array_dimensions() {
-        let array1d_1 =
-            Arc::new(ListArray::from_iter_primitive::<Int64Type, _, _>(vec![
-                Some(vec![Some(1), Some(2), Some(3)]),
-                Some(vec![Some(4), Some(5)]),
-            ]));
-        let array1d_2 =
-            Arc::new(ListArray::from_iter_primitive::<Int64Type, _, _>(vec![
-                Some(vec![Some(6), Some(7), Some(8)]),
-            ]));
-
-        let array2d_1 = Arc::new(array_into_list_array(array1d_1.clone())) as ArrayRef;
-        let array2d_2 = Arc::new(array_into_list_array(array1d_2.clone())) as ArrayRef;
-
-        let res = align_array_dimensions::<i32>(vec![
-            array1d_1.to_owned(),
-            array2d_2.to_owned(),
-        ])
-        .unwrap();
-
-        let expected = as_list_array(&array2d_1).unwrap();
-        let expected_dim = datafusion_common::utils::list_ndims(array2d_1.data_type());
-        assert_ne!(as_list_array(&res[0]).unwrap(), expected);
-        assert_eq!(
-            datafusion_common::utils::list_ndims(res[0].data_type()),
-            expected_dim
-        );
-
-        let array3d_1 = Arc::new(array_into_list_array(array2d_1)) as ArrayRef;
-        let array3d_2 = array_into_list_array(array2d_2.to_owned());
-        let res =
-            align_array_dimensions::<i32>(vec![array1d_1, Arc::new(array3d_2.clone())])
-                .unwrap();
-
-        let expected = as_list_array(&array3d_1).unwrap();
-        let expected_dim = datafusion_common::utils::list_ndims(array3d_1.data_type());
-        assert_ne!(as_list_array(&res[0]).unwrap(), expected);
-        assert_eq!(
-            datafusion_common::utils::list_ndims(res[0].data_type()),
-            expected_dim
-        );
-    }
 }
